Bug #33224: stream output of UDM CLI
Stream output instead of collecting it first as a list of lines and then
transferring them.

This improves the responsibility with large outputs and simplifies
debuggability, as previous output is no longer lost and a complete stack
trace is printed.

Throw exception in case of an error so simplify error handling: The CLI
front-end process expects the exact string "OPERATION FAILED" in the
output to terminate the process with an exit status unequal zero, which
was sprinkled over a lot of different places.