Bug #55141: ext-windows: code-block line continuation
The docs use the sphinx-copybutton extension that adds a little “copy”
button to the right of code-blocks. It is a lightweight wrapper around
the Javascript library ClipboardJS.

Readers read Univention documention as HTML and PDF. The copybutton
configuration caused problems with line continuation. It worked in HTML,
but not PDF, where it included characters not needed for copy.

This change defines `\` as line continuation character. Copybutton
recognizes lines after `\` as continued line. This doesn't need the `> `
on the continued line. It's easer to read in HTML and PDF and easier to
copy.