Bug #52230: Create user with password
When pwdLastSet is 0, the object is now resynced even if only
attributes on the irrelevant list are changed to catch changed passwords,
while the password is kept expired.
The test created a user without a password (and therefore with pwdLastSet=0)
to check that the object is NOT resynced if only attributes from the irrelevant
list are changed.