Bug #53885: remove UNIX socket before startup
The start fails because /run/univention-management-console/server.socket from the previous instance is left behind and thus the listen() fails.
Every 2nd restart doing a systemctl restart univention-management-console-server.service fails.

This is because of the broken code in server.py: If the UNIX socket exists it is deleted, so next time it will succeed. The correct way is to always delete the UNIX socket before the bind().