Bug #55177: Pipeline: Add CHANGELOG_TARGET_VERSION
For release notes and changelog we need the complete version number for
the deployment path. The pipeline variable `CHANGELOG_TARGET_VERSION`
serves this purpose. It makes sure that the documents `changelog` and
`release-notes` are published below `changelog/5.0-2` and
`release-notes/5.0-2`.

Update the doc/README.md with a hint on the document versioning.