fix(UCR): Atomic cache update
If some other process accesses the UCR cache for `.info` files in
`/var/cache/univention-config/cache` while an update is in progress, the
cache might be invalid, which triggers an automatic repair. This might
fail if some files are missing.

Write new cache to temporary file and rename only on success.

Bug #55230