Bug #54527: fix generate_junk_license() for Python 3
Traceback (most recent call last):
  File "80_udm_license", line 453, in <module>
    sys.exit(TestUMC.main())
  File "80_udm_license", line 446, in main
    self.run_junk_license_checks()
  File "80_udm_license", line 316, in run_junk_license_checks
    self.import_new_license(self.temp_license_folder + '/JunkTest.license')
  File "80_udm_license", line 157, in import_new_license
    assert request_result[0]["success"]
AssertionError

The uploaded license ended with '\nuniventionLicenseVirtualDesktopUsers: 0\nuniventionObjectType: settings/license\n\n/license\n\n'.
This is a malformed LDIF and was reported as:
`[{'message': 'Lizenzimport fehlgeschlagen: LDIF nicht wohlgeformt.', 'success': False}],`

The problem is that the replacement of the line for the univentionLicenseSignature replaced/inserted a string which is smaller than the original so that after writing the file the old content is also part of it.
Truncating the rest of the file content should fix it.