fix[udm]: mark default value for primaryGroup as changed in open() for not existing objects
otherwise the user is not added to the primaryGroup as no changes are
detected.
We could also simply remove the `self._set_default_group()` call but
then the user creation template for the UDM REST API would contain None
instead of the default primary group. This works, too, but for now it's
not a behavior change.

Bug #42080