refactor[UDM]: copy template default value explicitly
There could be some cases, where UDM(?) code updates complex syntax things,
which breaks when the value is stored in an immutable tuple instead of a modifiable list.
But if they are list you have to be careful of reference sharing, as updating one object
might then update another instance as well.
Therefore maybe this code tried to do a naive clone to prevent this?

Bug #55279