perf(udm users/user)!: add UCR variable for backwards compatibility
Using the group memberships via `memberOf` adds all groups to the user
which he is assigned to, even if the reading user cannot read the
specific groups of if the memberships are no groups/group objects.

As there might be code in the wild which don't do error handling when
iterating over group memberships a UCR variable
`directory/manager/user/group-memberships-via-memberof` can be used to
restore the old behavior.

The variable is going to be remove in UCS 5.1.

Bug #55269