fix(nvc): Run full Twisted reactor in main thread
As vnc-automate forks `teserracts` processes it needs to handle SIGCHLD.
The reactor installed by `vncdotool.api` is using
`installSignalHandlers=False`, which then does not setup the signal
handle and does not get notified when those prcesses terminate. This
leads to `processEnded()` not getting called and `gatherResults()` will
hang waiting for those Zombie processes.

Run the installation method in a sub-thread and the reactor in the main
thread, which is required for `signal.set_wakeup_fd()` to work.

Bug #45426