Bug #55530: [pipeline] Clone complete Git repo for release notes document
Per default GitLab makes a shallow copy of the repository. To determine
the last change on a document based on the git commit, Sphinx needs a full checkout.

This commit activates the full checkout only for the release notes and
there only for the HTML job, because a complete checkout needs a lot
more time.