Bug #55550: [pipeline] Set GIT_STRATEGY to clone for release notes
The `sphinx_last_updated_by_git` extension for Sphinx uses the Git
history of a file to determine the timestamp for the last change and
then uses it for the last updated information in the HTML build of a
document.

The build needs a full copy of the repository. The fetch strategy only
updates the existing checkout. And the existing doesn't have enough
history for the build and warns with git too shallow messages.

Setting the strategy to `clone` retrieves the complete repository. See
https://docs.gitlab.com/ee/ci/runners/configure_runners.html#git-strategy