fix: proxy/https defaulting
to proxy/http unless explicitly set to the empty string.

A http-proxy must implement GET and HEAD
A https-proxy must implement CONNECT instead, which proxy.py does not
do.

Instead of implementing the fallback for https to http in many locations
set it in the default layer of UCR, which allows to disable proxy/https
by explicitly setting it to the empty string.

Bug #55561