Bug #55609: [pipeline] skip-doc skips docs build, even with errata changes
The pipeline has the variable `pipeline` configured. With the value
`skip-doc` it skips the `doc-pipeline` job and therefore all
documentation jobs. But with one exception: Changes to `doc/errata` and
similar created the doc-pipeline job anyway.

Therefore, merge requests with errata YAML files always triggered the
`doc-pipeline` even with `$pipeline=skip-doc`.

This change is about the skip of the `doc-pipeline` without exception,
even with changes to `doc/errata`.