Issue #1457: Add reference to advisory guide
Having the content inside the YAML files produces extra effort for each
developer during replacing the guiding questions.

The content therefore was removed from the YAML template files. A
README.md file was added that mentions the advisory guide and links to
the guiding questions in the dev on boarding.

Refs #1457