Bug #55247: disable tls v3 for radius
We are currently using freeradius version 3.0.17, which wasn't developed
with full tls v3 support. The behaviour for eap and tls v3 was still
undefined.
The problem is that our openssl version 1.1.1 does negotiate tls v3, so
we need to explicitly disable it in freeradius.

https://github.com/FreeRADIUS/freeradius-server/issues/2385

This commit adds a UCR variable: "freeradius/conf/tls_max_version" with
the default value "1.2".