fix(umc): restart during upgrade 5.0-3+572..583
debhelper 9 still stops the old server in prerm and re-starts it in
postinst. Stopping is not protected my masking, so the UMC web session
is lost during the upgrade.
This is fatal when installing errata from USS during system setup as UMC
will then be stuck in the browser and the update process is killed.

~~Update to modern debhelper compat 12 from Debian-10-Buster.~~
Not done as this requires changed to setup.py as distutils is
deprecated.

Add a fixup package to remove the broken code from
u-m-c-web-server.prerm.

Bug #55753