fix(test): 00/03_check_release_signature_do_not_ignore_me
Scenario `update-from-1.2-start-4.3-4` is using
tests/scenarios/update-testing/update-from-1.2-start-4.3-4.cfg:

> kvm_operating_system: Others
> kvm_architecture: i386
                    ^^^^
> kvm_template: 1.2masterinstallation
> kvm_ucsversion: 4.3-4

Guess why
<https://jenkins.knut.univention.de:8181/job/UCS-4.4/job/UCS-4.4-9/job/Update%20Tests/lastCompletedBuild/test_system=update-from-1.2-start-4.3-4/testReport/00_checks/03_check_release_signature_do_not_ignore_me/master/>
is failing for 339 builds:

> libfaketime (/usr/lib/x86_64-linux-gnu/faketime/libfaketime.so.1) not found
                        ^^^^^^
See <https://wiki.debian.org/Multiarch/Tuples>

Do not exit with 1 as this flags the test as an ERROR, which it should
be marked as SKIPPED as a pre-requisite is not met.

Do NOT use `faketime` which does something strange on top of plain
libfaketime: `apt-get update` always fails and complains about invalid
signatures:
> Mindestens eine ungültige Signatur wurde entdeckt.

Refactor code and cleanup shellcheck issues.

Bug #53528