feat(udn): univention-translog reindex
Implement `univention-translog reinex` to re-built the transaction index
file in case it gets corrupted. Univention Directory Notifier already has
code to do maintain the index, but after certain error cases the index may
become corrupt and has to be re-built. The code in UDN is not optimized to
re-index many transactions in batch and shows performance issues for large
transaction files.

The real fix is to remove the index completely as it is no longer
required since protocol version 3, which is Bug #54837.

Bug #55863: univention-directory-notifier 13.0.3-11