univention-ssh: adjust to openssh change
With additional SSH_ASKPASS_REQUIRE=force, it works as of openssh-8.4-p1
(this version is already in Debian 11 Bullseye; it should therefore no longer work there either).
Since then, the default has changed: use_askpass = 0, allow_askpass = 0
and you have to use either prefer or force if STDIN is not a TTY; which it is not, because rsync needs that for communication with ssh.
This also means that the DISPLAY=:99 ... setsid hack can be omitted.

Bug #56593