fix(pam_runasroot): fix setting of environment variables due to sudo API change
A change in sudo (from version 1.9.4, commit 7d0b19d2a0236ff50f8e8253ba86e1672399ae52) causes
that when invoking this pam module, when setenv is used, the target for it is the private sudo environment
instead of the local environ pointer.

For that reason, when a program is invoked with execl, the state of the environment variables in the
target program is the original one, without the setenv operation having any apparent effect.

Because the environment modifications used in pam_runasroot are used only for the invoked processes
and do not need to be preserved afterwards, the execle function can be used together with a local pointer for the environment variables.

Bug #56579