DROPME: Flush postfix queue and dovecot auth cache
For some currently unknown reason, the dovecot's auth cache contains
wrong information. The test mails are correctly delivered to
/var/spool/dovecot/private///Maildir/....
via SMTP and LMTP into the IMAP mailbox.
But when ucs-test is accessing the mailbox via IMAP, the user
information seem to be invalid/unknown/not yet correct, so dovecot
is looking for the requested mailbox in
/var/spool/dovecot/private//...
which is obviously incorrect. This commit fixes the failure of the
test but not the actual problem in/with dovecot.

Bug #56648