fix(udm hook): fix idempotency and Python 3 compatibility of AttributeHook
map (UDM to LDAP) means to encode strings to bytes.
unmap (LDAP to UDM) means to decode list of bytes to strings.

When open() was called twice the current implementation broke. The
UCS@school importer for example makes two open() calls.

This is because hook_open() operates on the UDM values instead of the
LDAP values, which were already modified by the first open().

Bug #56036