refactor(USS): Profile variable reading
instead of having two functions `is_variable_set` and `get_profile_var`
to check for and read the profile variables, deprecate the 1st and
extend the 2nd to return the presence of the setting via a return code.

Please also notice that `is_variable_set` has a inverted logic and
returns `0` (success) it the variable is *not* set, which makes the
function awkward in an `if` statement.

Bug #34106