fix(USS): Re-create SSL on every change
the previous code did not work when a profile variable `ssl/*` did set a
setting to empty.

Collect all changes settings in an array and set them at once.

Directly work on files containing a dot instead of implementing the
filter manually.

Refactor parsing the new FQHN.

Remove comments not providing the "why".

Only create the symbolic link if it belongs to the same domain.

Bug #34106