Bug #27547: ignore tracebacks for importer test
Issue ucsschool#1031

In ucs-test there is a test, `99_end/01_var_log_tracebacks.py` that
checks to see if there are any tracebacks in the log file, and if so,
fails. We found that `90_ucsschool/302_http-api_class_column_no_school`
logs ValidationError tracebacks, but these tracebacks are an integral
part of the test, so should be ignored by the ucs-test.

In the `302_http-api_class_column_no_school` test we have [changed the
name of the OU](https://git.knut.univention.de/univention/ucsschool/-/merge_requests/239)
so that it is more distinguishable in the logs. This commit adds an
exclusion for the tracebacks that result.

Please note that this traceback has been occuring for over 168 runs, and
`git blame` and similar git tools have failed to identify where the
issue was introduced. For this reason, there isn't a bug number included
with the exclusions.