test(utils/set_repository_if_testing)
Rename from `check_repository_to_testing` to
`set_repository_if_testing`: it does not *check* anything, but actually
*changes* the repository server

The comment stated, that is must be called before `basic_setup`, but it
was called before that in many cases.

Consistently is upper-case variant `[ENV:RELEASE_UPDATE]` in two more
cases.

Bug #45426