doc: cross-link to local build
Building our documentation has a chicken-egg-issue: due to cross-linking
documents must and can only be built in the correct order.

This is made worse by always pulling in the "intersphnx" `objects.inv`
from the public web site, where the files are only published to at the
very end.

<https://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html>
allows specifying multiple locations, which are checked in turn. Add the
local build.

Bug #55696