fix(umc): make responding in non-main-thread thread safe
The only Tornado method that is thread-safe is `add_callback()`.
The consequences of calling `self.finish()` in the background thread are undefined:
sometimes it'll work, sometimes it'll be ignored, sometimes it raises an exception, sometimes it hangs forever, sometimes it does something else.
The correct approach therefor is using `self.ioloop.add_callback_reply, body)`.

Bug #43633