fix(replication): Allow ppolicy OIDs
In UCS 5.0 with OpenLDAP 2.4.47 `/etc/ldap/schema/ppolicy.schema` is
included, which defines 17 ATs and 2 OCs.
With UCS 5.2 and OpenLDAP 2.5 they are included with with the
`ppolicy.so` overlay module itself.

For UCS 5.0-4 they must *not* be excluded as otherwise OC `pwdPolicy`
will be missing and `cn=default,cn=ppolicy` cannot not be replicated.

Bug #55927