fix(replication): Allow ppolicy OIDs 3
In UCS 5.0 with OpenLDAP 2.4.47 `/etc/ldap/schema/ppolicy.schema` is
included, which defines 17 ATs and 2 OCs.
With UCS 5.2 and OpenLDAP 2.5 they are included with with the
`ppolicy.so` overlay module itself.

For UCS 5.0-4 they must *not* be excluded as otherwise OC `pwdPolicy`
will be missing and `cn=default,cn=ppolicy` cannot not be replicated.

Actually remove the lines as the `grep` from `debian/rules` otherwise
will still find the OIDs:
    $ grep --only '\<[012]\+\([.][0-9]\+\)\{3,\}\>' oid_skip.txt | wc -l
    954
    $ sed -ne 's/#.*//;s/\s\+/\n/;s/\n$//m;/./p' oid_skip.txt | wc -l
    935

Explanation:
1. remove comments starting with '#'
2. breaks into multiple lines on consecutive blanks
3. remove empty lines
4. print non-empty pattern space

Bug #55927