fix(umc-join): run join with more permissive umask
univention-join executed with umask 0077 (on UMC)

This results in all joinscripts also being run with umask 0077 and can result in wrong file permissions. For example:

-rw------- 1 root nogroup 1,9K Jun 10 14:04 /usr/lib/python3/dist-packages/univention/admin/syntax.d/app_syntax.py

The syntax is now only readable for root.

This has been previously discussed in some way or another: bug 21587, bug 25287 (bug 30112 broke it again), bug 33241 and possibly more.

How to reproduce on a replicant:
A) join using the umc
B)
- delete "/usr/lib/python3/dist-packages/univention/admin/syntax.d/app_syntax.py"
- force join script execution of 03univention-directory-listener using the umc

Bug #53431