Bug #27547: increase replication timeout for ucsshool env
On the UCS@school Multiserver Large env we create over 100,000 users in
a short amount of time. The listener stops running when the disk fills
up due to S4, and then doesn't start running again until S4 catches up.
This has caused test failures now for over a month.

We have several potential solutions to this problem:

* We get a bigger disk image.
* We set up some kind of pre-loaded LDAP so we don't need to worry about
  replication.
* We increase the `wait_for_replication` time.

The first two options are time-consuming and outside the scope of a
normal UCS@school test team ticket.

After monitoring the tests on the server, we don't suspect any loss of
data, as the S4 folder continues growing and the notifier keeps running,
despite the disk being reported as full for the listener. We do know
that if we wait long enough, the listener does eventually catch up.

So for a quick experiment, we would like to try increasing the
`wait_for_replication` time, to see if this gets the tests running
again, before we try other solutions.