fix(UMC): properly cancel aborted requests
Add a new route in the modserver that will cancel a request with the
given request id. A request ID is now generated whenever the UMC
makes a requests to the module server. This request ID replaces the old
request ID used in the module server.

When a UMC request which is cancelled from the client is detected a
HTTP request to "/cancel" will be send from the UMC server to the
module server containing the original request ID. This is is then
subsequently removed from the modules active request list, allowing it
to shutdown when signaled to.

To make sure the module is not getting killed when there are long
running threads, a new `running_threads` variable was added to
`SimpleThreads`. The module will only get killed if there are no running
threads left

Bug #56391