fix(umc): fix configurability of maximum request body size
The maximum request body size was configured in the UMC-Server but not
in the UMC module process and therefore lead to 400 bad requests.
Since the UMC rewrite file uploads aren't handeld anymore in the
UMC-Server core and given as short JSON information with a
dumped-filename to the module process but directly forwarded to the
module process this became an error.

Bug #56510