fix(vnc): Exception handling 2
Any exception is fatal for vncdotool.

Instead of using `VNCAutomateException` to handle the case where a text
is *not* visible, pass an additional `result` argument to receive the
location of the text and use that to decide for success or failure.

Bug #45426
Augments 7a1fd9d81ab836924791d9cfd2b25a67bc391bc0