fix(quota): fix aborted user query requests
Change the quota module implementation to use thread instead of non-blocking subprocesses.

Since git:9044299c48c2319f62584244fba3b50d4bea71a5 Bug #56391 requests are aborted / canceled when the connection was closed.
The module process ends itself when not active requests or running threads are there (after the module timeout).

Fetching the user quotas for a partition was done via non-blocking `tornado.Subprocess()` instead of a thread.
Therefore the module could just shut down after the connection was aborted.
On systems with many users (e.g. 3000) the quota/users/query wasn't answered within 10 minutes, which caused apache to close the connection.
The 10 minutes are not the umc/module/timeout (!) it also happens if that module timeout is below that value.
It also doesn't happen when we make requests directly against UMC server instead of the Apache gateway in front of it.
We don't know where the 10 minutes come from because the proxy-timeout we configured in apache is only 311 seconds.

The subprocess definately doesn't take 10 minutes but milliseconds to be executed:
$ time /usr/sbin/repquota -C -v /dev/mapper/vg_ucs-root
real    0m0,162s
user    0m0,018s
sys     0m0,033s

So another question remains, why is it okay for small systems but not for large systems?

Bug #56575