fix(udm/dns): IPv4/6 handling
Use `str.endswith((…))` to handle both IPv4 and IPv6.

`zoneName` is a required LDAP attribute using `IA5String`. Because of
this it is always set and can be decoded as ASCII.

host_record: Only prevent `in-addr.arpa` and `ip6.arpa` domains (Bug #55612).

ns_record: Allow in all zones as NS records are used for delegation,
which is allowed for all zones and all address families.

txt_record: Also allow for `ip6.arpa`.

Bug #50385
Bug #55612