fix(udm/dns): Fix check for DNS zones
`@` equals `$ORIGIN` in a BIND9 zone file as specified by RFC 1025
section 5.1. By using "-quoting and backslash-escaping *any* character /
octet can be used as a DNS name, even "@". As such it must not be used
to identify the zone record itself.

Instead of `sOARecord` as the identifier for zones.

Bug #50385