Bug #46120: rebase univention-antivir-mail UCR templates
UCS templates was really old and based on Debian 5. Merged
the templates with the ones from Debian 12. All the options in the
config file seem to be the default, so no functional changes should have
been introduced here.

The option `PidFile` has been removed from the upstream Debian config.
But as it still exists and is documented in `man clamd.conf` it is being
kept for backwards compatibility.