test(browser): Enable Chromium remote debugging
We now pass the `--remote-debugging-port` flag to Chromium which enables
us to connect to the headless browser using, for example, another
playwright instance. This should help with debugging stuck or otherwise
broken tests.

Before the `--remote-debugging-pipe` was already passed to Chromium,
however it was not possible to attach another Playwright instance.

Bug #56054