ci(doc): Fix artifacts path for _build directories
The important `out` directory worked, as expected and needed.

Nevertheless, the template for sphinx pipeline jobs suggests to also add
an artifact directory for the _build directory. The template doesn't
work for the UCS repository, because the build artifacts are stored in
`_build/$language/html` and not in `_build/html` as the template
expects. The pipeline jobs for Sphinx just omitted that directory and
showed a warning.

This commit now also adds the above mentioned _build directory to the
job artifacts for the Sphinx jobs. This is especially useful for
debugging of the jobs for spelling, linkcheck and building the PDF,
because those artifact directories contain log files. Of course, those
log files aren't needed for the actual output in the `out` directory.

Bug #56659