test(test/up/55): Fix failing component update
Most tests set `version/patchlevel=0`: Before UCS 5.0 starting from a
$minor > = or $patch > 0 would have requires generating all previous
releases as well as releases depend on all previous releases of the same
$major.
Starting with UCS 5.0-0 this is no longer necessary as each release is
complete by itself. Nevertheless the updates tests have never been
updated and still use the following hack:

The tests forcefully set `version/patchlevel=0` and then created
subsequent releases for testing.
But `mkpdi $major.$minor--componentt/...` has a bug as it uses the
*original* patch-level-version sorted in "$patchlevel": For
- 5.0-[01] it worked as $orig_pl == $patchlevel == 0, so 1 got created and
  updated to.
- 5.0-[2-] was broken: creating the component also created a release to
  which it was updated.

For 5.2-0 it is now broken again as there the original
$_version_patchlevel=0, but 1 gets created and updated to, failing the
comparison.

Change the code to create the component for patchlevel=0.

Bug #54868: Fix updater test for 5.0-2
Fixes: 4b657bac4482e6f4c022c9feac9e16563f24fd2e