DROPME: disable dovecot auth cache in tests
It looks like the IMAP client may be checking for new mails only during the auth.
We should have a look at dovecot auth code to check if this is true.

Reverts commit 7a0480b1b59d418cd2b2ed0da69cfbd2ed3dcfae
Bug #56648

---

DROPME: Flush postfix queue and dovecot auth cache

For some currently unknown reason, the dovecot's auth cache contains
wrong information. The test mails are correctly delivered to
/var/spool/dovecot/private///Maildir/....
via SMTP and LMTP into the IMAP mailbox.
But when ucs-test is accessing the mailbox via IMAP, the user
information seem to be invalid/unknown/not yet correct, so dovecot
is looking for the requested mailbox in
/var/spool/dovecot/private//...
which is obviously incorrect. This commit fixes the failure of the
test but not the actual problem in/with dovecot.

Bug #56648