refactor(appcenter): re-implement LooseVersion
The App Center depended on the LooseVersion class from the python3-distutils
module which is marked as deprecated since Python 3.10 and will no longer be
installed with Python 3.12. It is used for the comparison of version strings.
As many components still rely on it and we do not want to maintain deprecated
third-party code, we decided to re-implement all needed features.

LooseVersion types can be compared against string types and other LooseVersion
types. The NotImplemented singleton will be returned by its rich comparison
methods in case any other type is used for a comparison.

* python3-distutils dependencies were removed
* python3-distutils was removed from maintained-packages.txt
* LooseVersion class was re-implemented
* various examples for supported version comparisons were added to
  the docstrings.

Bug #55915