feat(udm-rest): add unused browser color scheme selection JS code
Nowadays browsers have a preference, wheather a dark or light theme
should be used.
We can query this, via JS or CSS selectors like: `@media (prefers-color-scheme: dark) { … }`.

We could include this JS snipped based on a configuration. It would
switch the default to the light theme, therefore don't enable it now.

Bug #56714