test(browser): move conftest.py into pytest plugin
To move more selenium tests to playwright and to reuse the fixtures and
functions of this library we need to make the conftest.py available to
other tests too.

Since we can't just move all the contents of the 86_browser/conftest.py
into the ucs-test top-level conftest.py a pytest plugin has been
created.

Bug #56054