fix(umc): remove on disk SAML identity cache
The on disk system DB SAML identity cache has been removed along with
the respective UCR variable to disable it. When adding an item to the
python shelve the underlying DBM sometimes throws an error for unknown
reasons.

At first we thought that it would be because of a parallel writing to
the DB from the multiple UMC processes in multiprocessing mode, however
this is not the case since every UMC multiprocessing processes is it's
own separately started python process.

Using the in-memory identity cache is fine since multiprocessing is
currently implemented using apache loadbalancing. Therefore requests of the same
session will always be routed the same instance of the UMC server.

Bug #54880