fix(univention.debug): fix univention.debug2 not sharing same handlers with univention.logging
Since the last commit which changed the ud loggers "MAIN" to
propagate=False the loggers from univention.debug which use the same
names don't propagate events to to the univention.debug2 handlers, which
are registered at the root logger instead of at the specific logger handlers.

Fixed by prefixing the univention.debug2 loggers with "ud2." and remove
the prefix in the log format.

Bug #55324