doc(udm/dns/zone/ns): Document FQDN
The zone name gets appended if no FQDN ending on a dot is given, e.g.
"ns" in "3.2.1.in-addr.arp" translates to "ns.3.2.1.in-addr.arpa.",
which is probably not what you want. As the server does not exist, the
zone will fail to load or might even prevent other zones form working.

Bug #50385