fix(umc): setting UCR policies in UMC no longer overrides old ones
UCR policies are multivalue, however this was not represented in the
UMC. However multiple UCR policies could be added via the UDM CLI. If
there was no UCR policy assigned to an object it was not possible to add
a new one via the UMC, since the `max` property of the `PolicyInput` was
set to 1.

However this would break when there were multiple UCR policies already
attached via the UDM CLI. The button to add a new policy is present,
however attempting to add a new UCR policy would lead to ONLY that new
policy being set and the old ones overwritten.

Bug #57046