doc(memberOf): updates and fixes
The attribute are not "presented" or "displayed", but "managed" by the
overlay.

The package `univention-ldap-overlay-memberof` is always installed as a
hard dependency of `univention-ldap-server` and no longer optional.

The module is enabled by default since UCS 4.3 and will be enforced with
UCS 5.2.

Mention UCRV to en-/disable module.

Mention matching UCRV to configure group class and attribute.

`univention-update-memberof` was already called on package installation
automatically, but this was pointless as the package is now a strong
dependency. But it is still required on manual activation.

Add section for troubleshooting.

Bug #33685
Closes: univention/ucs#2070