fix(udm): fix udm-cli zombie processes
[univention-cli-server:167](management/univention-directory-manager-modules/univention-cli-server#L167):

> rlist, _wlist, _xlist = select([sock], [], [], float(timeout))

If there are no new connections for the `socket` to accept, this will
wait until `timeout` expires. During that time no dead child processes
are reaped which is done by `ForkingMixIn.collect_children()`, which is
only called by `service_actions()` only *after* a request is handled.

Install a signal handler for `SIGCHLD` to overwrite the default
`SIG_IGN`, which will interrupt the `select()`.

Setup non-blocking in-process `pipe()` as a wakeup-FD to get an event
to interrupt `select()` on `SIGCHLD`.

Bug #7735