test(s4c): set timezone correctly in s4connector test
Previously in the s4connector test the timezone would be set with
'timedatectl set-timezone'. Before Debian 12/UCS5.2 this also changed the
'/etc/timezone' file, now it doesn't anymore.

This had the effect that the Python package 'tzlocal' didn't work
anymore. This package does a check if the contents of '/etc/timezone'
and the '/etc/localtime' symlink are matching, and if they don't it
raises an exception. This was now always the case since '/etc/timezone'
wasn't updated anymore by timedatectl.

Now set the timezone via the Debian mechanism of changing timezone,
which properly updates '/etc/timezone'.

Bug #56666