fix(umc): fix memory leaks / circular references
1. univention.management.console.resource.Resource.get_current_user()
assigns a fake translation function for the session, which holds a
reference to the Resource/RequestHandler. The RequestHandler is
therefore never garbage collected.

2. univention.management.console.session.Session passes "self" to .IACLs
and .Processes, which holds them so that a circular reference existed.
Note that Session.renew() throws away the IACLs and Processes instances,
without e.g. unassigning the .session reference, which will cause that
they never get garbage collected.

Bug #57104