fix(logging): detect correct univention-debug log level if
univention.debug is not initialized by univention.logging
For example in S4-Connector.

Since Bug #56970 we are importing "univention.logging" in univention/admin/__init__.py and univention/uldap.py before univention-debug is initialized.

This causes that the UDM univention-debug loglevel is not respected anymore for every log entry written by the univention.logging interface (which is now done for the whole UDM and UMC code). Affected components are e.g. S4-Connector and AD-Connector, etc.

Example with a order in which the correct level is detected:
```
>>> import univention.debug as ud
>>> ud.init('stdout', ud.FLUSH, ud.NO_FUNCTION)
29.02.24 19:17:07.846  DEBUG_INIT
<_io.FileIO name=1 mode='rb+' closefd=False>
>>> ud.set_level(ud.ADMIN, ud.ALL)
>>> import univention.logging
>>> A = univention.logging.getLogger('ADMIN')
>>> A
<univention.logging.Logger ADMIN (Level 9)>
>>> A.univention_debug_handler.get_ud_level()
4
>>> A.univention_debug_handler.getLevel()
9
>>> A.getEffectiveLevel()
9
```

Example with the other order, which doesn't detect the correct level:
```
>>> import univention.debug as ud
>>> import univention.logging
>>> import logging
>>> A = univention.logging.getLogger('ADMIN')
>>> A
<univention.logging.Logger ADMIN (ERROR)>
>>> ud.init('stdout', ud.FLUSH, ud.NO_FUNCTION)
29.02.24 19:14:17.109  DEBUG_INIT
<_io.FileIO name=1 mode='rb+' closefd=False>
>>> ud.set_level(ud.ADMIN, ud.ALL)
>>> A
<univention.logging.Logger ADMIN (ERROR)>
>>> A.info('test')  # nothing is logged :-(
>>> ud.debug(ud.ADMIN, ud.INFO, 'test')
29.02.24 19:26:41.940  ADMIN       ( INFO    ) : test
>>> A.univention_debug_handler.get_ud_level()  # correct
4
>>> A.univention_debug_handler.getLevel()  # correct
9
>>> A.getEffectiveLevel()  # incorrect
40
>>> logging.ERROR
40
```

Bug #57101