fix(umc): fix LDAP connection and memory leak
`univention.management.console.message.Request.get_user_ldap_connection()`
adds a `self.bind_user_connection` reference to a LDAP connection cache
(in `.ldap.LDAP.__ldap_connections`), which references the Request object
via self, so that this is never garbage collected.
Additionally the hash of that function is used as a cache indicator.
Prior to UMC unification the method was not part of the Request object
but the long living Instance base class of the module. Therefor for each
request a new LDAP connection is opened.

Bug #57113