fix(playwright): only load necessary pytest modules
in `git:9a2580896787c0589e50f` python3-pytest-asyncio was added as a
ucs-test dependency. Due to the /usr/share/ucs-test/playwright-venv
virtual environment being created with `--system-site-packages` this
pytest plugin was picked up by the venv. However the pytest version
installed inside the venv is too new to work with the pytest-asyncio
plugin. This led to the playwright tests not starting anymore.

Bug #56054