feat(umc): Cookie creation now also specifies the domain
The problem was that univention-web did not specify a domain when creating cookies,
whereas univention-portal did. For example, if the domain is portal-dev.intranet,
the portal has created a cookie with the domain ".portal-dev.intranet".
However, the parameter was omitted for univention-web. As a result,
"portal-dev.intranet" was used as the domain. This resulted in us having two cookies
with the same name in `document.cookies` and the code took the first cookie in the string.
Now that we are using the same domain, we no longer have this duplication and the language remains consistent.

Bug #57192