fix(bibliography): Update link to Debian installation guide
The bibliography entry refers to the preseed appendix of the Debian Installation Guide.
The Guide is only available for the stable and oldstable version of Debian.
For UCS 5.0, we'd need the oldoldstable "Buster" version, which is no longer available.

I changed the link to the oldstable version. It doesn't exactly match
the UCS 5.0 Debian base, but it's closer to it than the Debian stable
version.

Issue #2242