fix(dev-reference): Example using deprecated repository/online/component../prefix
The variable repository/online/component/NAME/prefix is deprecated and
was used in an example. I replaced the example with the recommended
repository/online/component/NAME/server.

I also changed the value for repository/online/component/NAME from `yes`
to `enabled` as mentioned in the reference.

Bug #57238
Issue univention/ucs#2225