fix(self-service): Use sambaPwdLastSet in invitation mail listener filter
The attributes `shadowMax`/`shadowLastChange` are influenced
by password policies. When a user is created and a password policy is
applied, this leads to a non-matching self-service filter and as a
result no email is sent. `sambaPwdLastSet` is 0 when a user is created
and pwdChangeNextLogin has been set to 1 and is now used in the filter.

Bug #57226